requirements.txt
pyyaml>=6.0.1
requests>=2.31.0
chromadb>=0.4.0
streamlit>=1.28.0
tqdm>=4.66.0

